requirements.txt
Python==3.6.4

asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
Django==2.0.7
django-rest-knox==3.1.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
idna==2.7
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.5
six==1.11.0